begin vg call refactor
